Trait manual_xcm_rpc::ManualXcmApi

pub trait ManualXcmApi: Sized + Send + Sync + 'static {
    fn inject_downward_message(
        &self, 
        message: Vec<u8>
    ) -> BoxFuture<'static, RpcResult<()>>;

    fn inject_hrmp_message(
        &self, 
        sender: ParaId, 
        message: Vec<u8>
    ) -> BoxFuture<'static, RpcResult<()>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }
}

This RPC interface is used to manually submit XCM messages that will be injected into a parachain-enabled runtime. This allows testing XCM logic in a controlled way in integration tests.

Required methods

fn inject_downward_message(
&self,
message: Vec<u8>
) -> BoxFuture<'static, RpcResult<()>>

Inject a downward xcm message - A message that comes from the relay chain. You may provide an arbitrary message, or if you provide an emtpy byte array, Then a default message (DOT transfer down to ALITH) will be injected

fn inject_hrmp_message(
&self,
sender: ParaId,
message: Vec<u8>
) -> BoxFuture<'static, RpcResult<()>>

Inject an HRMP message - A message that comes from a dedicated channel to a sibling parachain.

Cumulus Parachain System seems to have a constraint that at most one hrmp message will be sent on a channel per block. At least that’s what this comment implies: https://github.com/paritytech/cumulus/blob/c308c01b/pallets/parachain-system/src/lib.rs#L204 Neither this RPC, nor the mock inherent data provider make any attempt to enforce this constraint. In fact, violating it may be useful for testing. The method accepts a sending paraId and a bytearray representing an arbitrary message as parameters. If you provide an emtpy byte array, then a default message representing a transfer of the sending paraId’s native token will be injected.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.